increased wait times, fixed some bugs in tests
for the labels, appearently, constants are needed instead of arbitrary strings. That makes perfect sense, though, because some of the labels can be localized. Additionally, Swift seems to have a problem with scopes, at least in the debugger. Thus, I changed `if let person = person` to `if let person = optionalPerson` to avoid confusion